feat(ui): 2D default graph, lazy 3D, progressive loading
Phase 1: Switch to 2D force-directed graph as default using Canvas-based
force-graph library. 2D/3D toggle in toolbar. Labels always readable,
no camera rotation needed, dramatically lighter on CPU/GPU.

Phase 2: Three.js + 3d-force-graph lazy-loaded via dynamic import only
when user toggles to 3D mode. Default graph tab loads ~90KB instead of
~1.3MB. Three.js bundle split into separate chunk.

Phase 3: Progressive loading — initial render shows top 20 nodes by
pagerank (<500ms), then background-fetches full 200 nodes. Force
simulation preserves existing node positions when adding new data.

Bundle impact:
- Default (2D): GraphView 11KB + force-graph 79KB
- 3D (lazy): Graph3D 5KB + graph3d 1,307KB (only loaded on toggle)
- Main bundle: 824KB (Three.js no longer included)
