+++ ui/package.json
@@ -37,6 +37,7 @@
     "3d-force-graph": "^1.79.1",
     "codemirror": "^6.0.2",
     "dompurify": "^3.2.0",
+    "force-graph": "^1.51.1",
     "highlight.js": "^11.11.0",
     "marked": "^15.0.0",
     "three": "^0.182.0"
